Changes proposed in pull request comments
