+++ build.gradle
@@ -7,9 +7,7 @@
 apply plugin: 'java'
 apply plugin: 'application'
 
-//mainClassName = 'MonitorLauncher'
-mainClassName = 'edu.cmu.sei.kalki.IotInterface'
-//mainClassName = 'edu.cmu.sei.kalki.Mail.SendMail'
+mainClassName = 'edu.cmu.sei.kalki.iotinterface.IotInterface'
 
 sourceCompatibility = 1.8
 
@@ -22,9 +20,6 @@ repositories {
 }
 
 dependencies {
-//    compile name: 'slf4j-api-1.6.1'
-//    compile name: 'mail'
-//    compile name: 'activation'
     compile group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '9.4.19.v20190610'
     compile group: 'org.subethamail', name: 'subethasmtp', version: '3.1.3'
     compile 'org.json:json:20171018'
@@ -33,5 +28,4 @@ dependencies {
     compile name: 'huesdkresources'
     compile 'com.jcraft:jsch:0.1.55'
     compile 'org.apache.httpcomponents:httpclient:4.5.10'
-//    compile name: 'tablelayout'
 }
