+++ src/main/java/edu/cmu/sei/kalki/iotinterface/utils/mail/EventObserver.java
@@ -1,0 +1,5 @@
+package edu.cmu.sei.kalki.iotinterface.utils.mail;
+
+public interface EventObserver {
+    public void notify(String message);
+}
