--- src/main/java/edu/cmu/sei/kalki/utils/mail/EventObserver.java
@@ -1,5 +1,0 @@
-package edu.cmu.sei.kalki.utils.mail;
-
-public interface EventObserver {
-    public void notify(String message);
-}
