--- src/main/java/edu/cmu/sei/kalki/utils/mail/SendMail.java
@@ -1,51 +1,0 @@
-package edu.cmu.sei.kalki.utils.mail;
-
-import java.util.*;
-import javax.mail.*;
-import javax.mail.internet.*;
-
-public class SendMail {
-
-    public static void main(String [] args) {
-        // Recipient's email ID needs to be mentioned.
-        String to = "test@gmail.com";
-
-        // Sender's email ID needs to be mentioned
-        String from = "camera1@dlink.com";
-
-        // Assuming you are sending email from localhost
-        String host = "localhost";
-
-        // Get system properties
-        Properties properties = System.getProperties();
-
-        // Setup mail server
-        properties.setProperty("mail.smtp.host", host);
-
-        // Get the default Session object.
-        Session session = Session.getDefaultInstance(properties);
-
-        try {
-            // Create a default MimeMessage object.
-            MimeMessage message = new MimeMessage(session);
-
-            // Set From: header field of the header.
-            message.setFrom(new InternetAddress(from));
-
-            // Set To: header field of the header.
-            message.addRecipient(Message.RecipientType.TO, new InternetAddress(to));
-
-            // Set Subject: header field
-            message.setSubject("This is the Subject Line!");
-
-            // Now set the actual message
-            message.setText("This is actual message");
-
-            // Send message
-            Transport.send(message);
-            System.out.println("Sent message successfully....");
-        } catch (MessagingException mex) {
-            mex.printStackTrace();
-        }
-    }
-}
